# app/dist/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "0.0.0",
        "title": "User API"
    },
    "host": "localhost:3000",
    "basePath": "/v1",
    "paths": {
        "/users": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get Users",
                "description": "return all users\n",
                "parameters": [
                    {
                        "name": "x-access-token",
                        "in": "header",
                        "description": "User Access Token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Array of User Object",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/User"
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Register User",
                "description": "The Users endpoint returns information about the user\n",
                "parameters": [
                    {
                        "name": "firstName",
                        "in": "formData",
                        "description": "First name of User",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "lastName",
                        "in": "formData",
                        "description": "Last Name of User",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "email",
                        "in": "formData",
                        "description": "Email of User",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "formData",
                        "description": "Password of User",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "type",
                        "in": "formData",
                        "description": "Role of User",
                        "required": true,
                        "type": "string",
                        "enum": [
                            "USER"
                        ]
                    }
                ],
                "responses": {
                    "200": {
                        "description": "An Object of User",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/user/{userId}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get User",
                "description": "This endpoint returns information about the user by its id.\n",
                "parameters": [
                    {
                        "name": "x-access-token",
                        "in": "header",
                        "description": "User Access Token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "Id of User",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "An Object of User",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Update User",
                "description": "This endpoint updates information of the user by its id.\n",
                "parameters": [
                    {
                        "name": "x-access-token",
                        "in": "header",
                        "description": "User Access Token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "Id of User",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "firstName",
                        "in": "formData",
                        "description": "First name of User",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "lastName",
                        "in": "formData",
                        "description": "Last name of User",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "An Object of User",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/user/authenticate": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Authenticate user User",
                "description": "The Users endpoint returns information about the user and auth token\n",
                "parameters": [
                    {
                        "name": "email",
                        "in": "formData",
                        "description": "User Email",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "formData",
                        "description": "User Password",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "An Object of User",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/user/authorize/google": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Authorize User with google auth info.",
                "description": "The Users endpoint returns information about the user\n",
                "parameters": [
                    {
                        "name": "accessToken",
                        "in": "formData",
                        "description": "Google User Access Token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "type",
                        "in": "formData",
                        "description": "User Type",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "An Object of User",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/user/authorize/facebook": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Authorize User with facebook auth info.",
                "description": "The Users endpoint returns information about the user\n",
                "parameters": [
                    {
                        "name": "accessToken",
                        "in": "formData",
                        "description": "Facebook User Access Token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "type",
                        "in": "formData",
                        "description": "User Type",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "An Object of User",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "Error": {
            "type": "object",
            "properties": {
                "statusCode": {
                    "type": "integer",
                    "format": "int32"
                },
                "error": {
                    "type": "string"
                }
            }
        },
        "User": {
            "type": "object",
            "properties": {
                "firstName": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "date_created": {
                    "type": "string"
                }
            }
        }
    }
}